--- a/src/main/java/org/roaringbitmap/buffer/MappeableArrayContainer.java
+++ b/src/main/java/org/roaringbitmap/buffer/MappeableArrayContainer.java
@@ -748,7 +748,74 @@ public final class MappeableArrayContainer extends MappeableContainer implements
 
   @Override
   public MappeableContainer ior(final MappeableArrayContainer value2) {
-    return this.or(value2);
+    final int totalCardinality = getCardinality() + value2.getCardinality();
+    if (totalCardinality > DEFAULT_MAX_SIZE) {// it could be a bitmap!
+      final MappeableBitmapContainer bc = new MappeableBitmapContainer();
+      if (!BufferUtil.isBackedBySimpleArray(bc.bitmap)) {
+        throw new RuntimeException("Should not happen. Internal bug.");
+      }
+      long[] bitArray = bc.bitmap.array();
+      if (BufferUtil.isBackedBySimpleArray(value2.content)) {
+        short[] sarray = value2.content.array();
+        for (int k = 0; k < value2.cardinality; ++k) {
+          short v = sarray[k];
+          final int i = BufferUtil.toIntUnsigned(v) >>> 6;
+          bitArray[i] |= (1L << v);
+        }
+      } else {
+        for (int k = 0; k < value2.cardinality; ++k) {
+          short v2 = value2.content.get(k);
+          final int i = BufferUtil.toIntUnsigned(v2) >>> 6;
+          bitArray[i] |= (1L << v2);
+        }
+      }
+      if (BufferUtil.isBackedBySimpleArray(content)) {
+        short[] sarray = content.array();
+        for (int k = 0; k < cardinality; ++k) {
+          short v = sarray[k];
+          final int i = BufferUtil.toIntUnsigned(v) >>> 6;
+          bitArray[i] |= (1L << v);
+        }
+      } else {
+        for (int k = 0; k < cardinality; ++k) {
+          short v = content.get(k);
+          final int i = BufferUtil.toIntUnsigned(v) >>> 6;
+          bitArray[i] |= (1L << v);
+        }
+      }
+      bc.cardinality = 0;
+      int len = bc.bitmap.limit();
+      for (int index = 0; index < len; ++index) {
+        bc.cardinality += Long.bitCount(bitArray[index]);
+      }
+      if (bc.cardinality <= DEFAULT_MAX_SIZE) {
+        return bc.toArrayContainer();
+      } else if (bc.isFull()) {
+        return MappeableRunContainer.full();
+      }
+      return bc;
+    }
+    if (totalCardinality >= content.limit()) {
+      increaseCapacity(totalCardinality);
+    }
+    BufferUtil.arraycopy(content, 0, content, value2.cardinality, cardinality);
+    if (BufferUtil.isBackedBySimpleArray(content)
+            && BufferUtil.isBackedBySimpleArray(value2.content)) {
+      cardinality =
+              Util.unsignedUnion2by2(
+                      content.array(), value2.cardinality, cardinality,
+                      value2.content.array(), 0, value2.cardinality,
+                      content.array()
+              );
+    } else {
+      cardinality =
+              BufferUtil.unsignedUnion2by2(
+                      content, value2.cardinality, cardinality,
+                      value2.content, 0, value2.cardinality,
+                      content.array()
+              );
+    }
+    return this;
   }
 
   @Override
@@ -1042,11 +1109,18 @@ public final class MappeableArrayContainer extends MappeableContainer implements
     if (BufferUtil.isBackedBySimpleArray(value1.content)
         && BufferUtil.isBackedBySimpleArray(value2.content)) {
       answer.cardinality =
-          org.roaringbitmap.Util.unsignedUnion2by2(value1.content.array(), value1.getCardinality(),
-              value2.content.array(), value2.getCardinality(), answer.content.array());
+          Util.unsignedUnion2by2(
+                  value1.content.array(), 0, value1.getCardinality(),
+                  value2.content.array(), 0, value2.getCardinality(),
+                  answer.content.array()
+          );
     } else {
-      answer.cardinality = BufferUtil.unsignedUnion2by2(value1.content, value1.getCardinality(),
-          value2.content, value2.getCardinality(), answer.content.array());
+      answer.cardinality =
+              BufferUtil.unsignedUnion2by2(
+                      value1.content, 0, value1.getCardinality(),
+                      value2.content, 0, value2.getCardinality(),
+                      answer.content.array()
+              );
     }
     return answer;
   }
@@ -1095,11 +1169,18 @@ public final class MappeableArrayContainer extends MappeableContainer implements
     if (BufferUtil.isBackedBySimpleArray(value1.content)
         && BufferUtil.isBackedBySimpleArray(value2.content)) {
       answer.cardinality =
-          org.roaringbitmap.Util.unsignedUnion2by2(value1.content.array(), value1.getCardinality(),
-              value2.content.array(), value2.getCardinality(), answer.content.array());
+              Util.unsignedUnion2by2(
+                      value1.content.array(), 0, value1.getCardinality(),
+                      value2.content.array(), 0, value2.getCardinality(),
+                      answer.content.array()
+              );
     } else {
-      answer.cardinality = BufferUtil.unsignedUnion2by2(value1.content, value1.getCardinality(),
-          value2.content, value2.getCardinality(), answer.content.array());
+      answer.cardinality =
+              BufferUtil.unsignedUnion2by2(
+                      value1.content, 0, value1.getCardinality(),
+                      value2.content, 0, value2.getCardinality(),
+                      answer.content.array()
+              );
     }
     return answer;
   }