I stumbled on this operator while working on my personal dataset. The proposed changes simply leverage an optimization that has been used before: if the backed array is large enough to hold the result and its initial content, push the content at the end of the array and write the result at the beginning. 
With less pressure on the garbage collector, I noticed a ~20% improvement in my usecase.
To validate the changes, I added a benchmark on our real dataset. See below the results. There is always either a speedup or no significant regression.

| Benchmark                        | (dataset)              | (immutable) | (type)           | Score before | Score after | Variation |
|----------------------------------|------------------------|-------------|------------------|--------------|-------------|-----------|
| RealDataBenchmarkIOr.pairwiseIOr | census-income          | FALSE       | roaring          | 2435.091     | 2285.147    | 6.16%     |
| RealDataBenchmarkIOr.pairwiseIOr | census-income          | FALSE       | roaring_with_run | 276.821      | 270.427     | 2.31%     |
| RealDataBenchmarkIOr.pairwiseIOr | census1881             | FALSE       | roaring          | 5985.871     | 6116.565    | -2.18%    |
| RealDataBenchmarkIOr.pairwiseIOr | census1881             | FALSE       | roaring_with_run | 7772.466     | 7437.919    | 4.30%     |
| RealDataBenchmarkIOr.pairwiseIOr | dimension_008          | FALSE       | roaring          | 2402.474     | 2364.973    | 1.56%     |
| RealDataBenchmarkIOr.pairwiseIOr | dimension_008          | FALSE       | roaring_with_run | 4068.806     | 3970.744    | 2.41%     |
| RealDataBenchmarkIOr.pairwiseIOr | dimension_003          | FALSE       | roaring          | 15017.826    | 15273.254   | -1.70%    |
| RealDataBenchmarkIOr.pairwiseIOr | dimension_003          | FALSE       | roaring_with_run | 7748.019     | 6580.877    | 15.06%    |
| RealDataBenchmarkIOr.pairwiseIOr | dimension_033          | FALSE       | roaring          | 2674.554     | 2706.902    | -1.21%    |
| RealDataBenchmarkIOr.pairwiseIOr | dimension_033          | FALSE       | roaring_with_run | 1176.172     | 1197.942    | -1.85%    |
| RealDataBenchmarkIOr.pairwiseIOr | uscensus2000           | FALSE       | roaring          | 418.675      | 420.828     | -0.51%    |
| RealDataBenchmarkIOr.pairwiseIOr | uscensus2000           | FALSE       | roaring_with_run | 437.241      | 408.776     | 6.51%     |
| RealDataBenchmarkIOr.pairwiseIOr | weather_sept_85        | FALSE       | roaring          | 1914.518     | 1899.064    | 0.81%     |
| RealDataBenchmarkIOr.pairwiseIOr | weather_sept_85        | FALSE       | roaring_with_run | 1989.191     | 1914.704    | 3.74%     |
| RealDataBenchmarkIOr.pairwiseIOr | wikileaks-noquotes     | FALSE       | roaring          | 5180.151     | 5092.232    | 1.70%     |
| RealDataBenchmarkIOr.pairwiseIOr | wikileaks-noquotes     | FALSE       | roaring_with_run | 13735.464    | 12436.016   | 9.46%     |
| RealDataBenchmarkIOr.pairwiseIOr | census-income_srt      | FALSE       | roaring          | 1189.468     | 1133.497    | 4.71%     |
| RealDataBenchmarkIOr.pairwiseIOr | census-income_srt      | FALSE       | roaring_with_run | 81.35        | 70.371      | 13.50%    |
| RealDataBenchmarkIOr.pairwiseIOr | census1881_srt         | FALSE       | roaring          | 8143.423     | 7826.395    | 3.89%     |
| RealDataBenchmarkIOr.pairwiseIOr | census1881_srt         | FALSE       | roaring_with_run | 9015.314     | 9221.516    | -2.29%    |
| RealDataBenchmarkIOr.pairwiseIOr | weather_sept_85_srt    | FALSE       | roaring          | 702.266      | 706.214     | -0.56%    |
| RealDataBenchmarkIOr.pairwiseIOr | weather_sept_85_srt    | FALSE       | roaring_with_run | 1081.853     | 1046.578    | 3.26%     |
| RealDataBenchmarkIOr.pairwiseIOr | wikileaks-noquotes_srt | FALSE       | roaring          | 5601.957     | 5218.003    | 6.85%     |
| RealDataBenchmarkIOr.pairwiseIOr | wikileaks-noquotes_srt | FALSE       | roaring_with_run | 2782.501     | 2801.871    | -0.70%    |