@lemire 2017-04-07
@gssiyankai The code is straight-forward and it seems clear that it cannot hurt (much). I take it you recommend we push this in a release?