@gssiyankai 2017-04-07
@lemire Indeed.